Majority of the files converted to tsx and basic ts errors fixed.
