+++ src/components/NotFound.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+/* 404 page */
+
+import {
+    Box
+} from '@chakra-ui/react'
+
+const NotFound = () => {
+    return (
+        <Box mt={4} px={24}>
+            <h3>Nope, not here. 404.</h3>
+        </Box>
+    )
+}
+
+export default NotFound;
