+++ src/components/UserProfile.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import Authentication from './Authentication';
+
+/* Calls for authentication of user, maybe redundent */
+
+const UserProfile = ({ user, userLangPrefs, handleMessage }) => (
+    <div>
+        <Authentication user={user} userLangPrefs={userLangPrefs} handleMessage={handleMessage}/>
+    </div>
+)
+
+export default UserProfile;
