--- src/components/quotes.jsx
@@ -1,68 +1,0 @@
-export const quotes=[
-    {
-        quote:"Hello, world!",
-        author:"Brian Kernighan",
-        originalWord: "Hello,",
-        revealWord: "ಹಲೋ," 
-    },
-    {
-        quote:"There is no exquisite beauty without some strangeness in the proportion.",
-        author:"Francis Bacon",
-        originalWord: "strangeness",
-        revealWord: "奇妙さ" 
-    },
-    {
-        quote:"In memory everything seems to happen to music.",
-        author:"Tennessee Williams",
-        originalWord: "music.",
-        revealWord: "음악." 
-    },
-    {
-        quote:"Everybody needs his memories. They keep the wolf of insignificance from the door.",
-        author:"Saul Bellow",
-        originalWord: "wolf",
-        revealWord: "loup" 
-    },
-    {
-        quote:"It's a poor sort of memory that only works backwards.",
-        author:"Lewis Carroll",
-        originalWord: "backwards.",
-        revealWord: "visszafelé" 
-    },
-    {
-        quote:"Memory is not wisdom; idiots can rote volumes.",
-        author:"Martin Farquhar Tupper",
-        originalWord: "idiots",
-        revealWord: "aptallar" 
-    },
-    {
-        quote:"Blessed are the forgetful, for they get the better even of their blunders.",
-        author:"Friedrich Nietzsche",
-        originalWord: "forgetful,",
-        revealWord: "ขี้ลืม" 
-    },
-    {
-        quote:"Language is to the mind more than light is to the eye.",
-        author:"William Gibson",
-        originalWord: "Language",
-        revealWord: "Taal" 
-    },
-    {
-        quote:"Language is a virus from outer space",
-        author:"William S. Burroughs",
-        originalWord: "virus",
-        revealWord: "ιός" 
-    },
-    {
-        quote:"I personally believe we developed language because of our deep inner need to complain.",
-        author:"Jane Wagner",
-        originalWord: "complain.",
-        revealWord: "lamentarsi" 
-    },
-    {
-        quote:"I have had so many names, old names that only the wind and trees can pronounce.",
-        author:"el fauno",
-        originalWord: "wind",
-        revealWord: "viento" 
-    }
-]
