+++ src/components/speechLanguagesMap.tsx
@@ -1,0 +1,4 @@
+// hand crafted list of google translate codes to google tts codes
+// This is static and would take considerable more work to get to work dynamically
+// So, good for now
+export const speechLanguageMap = {"ar":{"language":"Arabic","ttsCode":"ar-XA"},"bn":{"language":"Bengali","ttsCode":"bn-IN"},"cs":{"language":"Czech","ttsCode":"cs-CZ"},"da":{"language":"Danish","ttsCode":"da-DK"},"de":{"language":"German","ttsCode":"de-DE"},"el":{"language":"Greek","ttsCode":"el-GR"},"en":{"language":"English","ttsCode":"en-UK"},"es":{"language":"Spanish","ttsCode":"es-ES"},"fr":{"language":"French","ttsCode":"fr-FR"},"gu":{"language":"Gujarati","ttsCode":"gu-IN"},"hi":{"language":"Hindi","ttsCode":"hi-IN"},"hu":{"language":"Hungarian","ttsCode":"hu-HU"},"id":{"language":"Indonesian","ttsCode":"id-ID"},"it":{"language":"Italian","ttsCode":"it-IT"},"ja":{"language":"Japanese","ttsCode":"ja-JP"},"kn":{"language":"Kannada","ttsCode":"kn-IN"},"ko":{"language":"Korean","ttsCode":"ko-KR"},"ml":{"language":"Malayalam","ttsCode":"ml-IN"},"nl":{"language":"Dutch","ttsCode":"nl-NL"},"no":{"language":"Norwegian","ttsCode":"nb-NO"},"pl":{"language":"Polish","ttsCode":"pl-PL"},"pt":{"language":"Portuguese","ttsCode":"pt-PT"},"ru":{"language":"Russian","ttsCode":"ru-RU"},"sk":{"language":"Slovak","ttsCode":"sk-SK"},"sv":{"language":"Swedish","ttsCode":"sv-SE"},"ta":{"language":"Tamil","ttsCode":"ta-IN"},"te":{"language":"Telugu","ttsCode":"te-IN"},"th":{"language":"Thai","ttsCode":"th-TH"},"tl":{"language":"Tagalog","ttsCode":"fil-PH"},"tr":{"language":"Turkish","ttsCode":"tr-TR"},"uk":{"language":"Ukrainian","ttsCode":"uk-UA"},"vi":{"language":"Vietnamese","ttsCode":"vi-VN"},"zh-CN":{"language":"Chinese","ttsCode":"cmn-CN"},"zh-TW":{"language":"Traditional chinese","ttsCode":"cmn-TW"}}
