+++ tsconfig.json
@@ -5,7 +5,7 @@
     "strictNullChecks": true, // enable strict null checks as a best practice
     "module": "commonjs", // specify module code generation
     "jsx": "react", // use typescript to transpile jsx to js
-    "target": "es5", // specify ECMAScript target version
+    "target": "es6", // specify ECMAScript target version
     "allowJs": true, // allow a partial TypeScript and JavaScript codebase
     "allowSyntheticDefaultImports": true,
     "moduleResolution": "node"
